move the Get command to its own package
